refactor: 랜덤한 수를 가지는 배열 만드는 함수 math.js에서 array.js로 분리
