+++ src/Domain/Model/LottoMachine.js
@@ -1,7 +1,7 @@
 import { LOTTO_DEFINITION } from '../Constant/definition.js';
 import Lotto from './Lotto.js';
 import { sortAscending } from '../../Utils/sorting.js';
-import { makeNotDuplicatedRandomNumbers } from '../../Utils/math.js';
+import { makeNotDuplicatedRandomNumbers } from '../../Utils/array.js';
 
 class LottoMachine {
   #lottoList;
@@ -10,18 +10,21 @@ class LottoMachine {
     return Math.floor(money / LOTTO_DEFINITION.ONE_PRICE);
   }
 
+  #makeLotto() {
+    const numbers = makeNotDuplicatedRandomNumbers(
+      LOTTO_DEFINITION.NUMBER_COUNTS,
+      {
+        min: LOTTO_DEFINITION.MIN_NUMBER,
+        max: LOTTO_DEFINITION.MAX_NUMBER,
+      },
+    );
+    return new Lotto(sortAscending(numbers));
+  }
+
   makeLottoList(lottoCount) {
-    this.#lottoList = Array.from({ length: lottoCount }, () => {
-      const numbers = makeNotDuplicatedRandomNumbers(
-        LOTTO_DEFINITION.NUMBER_COUNTS,
-        {
-          min: LOTTO_DEFINITION.MIN_NUMBER,
-          max: LOTTO_DEFINITION.MAX_NUMBER,
-        },
-      );
-      sortAscending(numbers);
-      return new Lotto(numbers);
-    });
+    this.#lottoList = Array.from({ length: lottoCount }, () =>
+      this.#makeLotto(),
+    );
   }
 
   getLottoList() {
